fix all pages
